Begin work on desktop display
