+++ src/App.js
@@ -12,39 +12,38 @@ import RollDicePrompt from 'roll-dice';
 import EventHistory from 'event/history-panel';
 
 import { GameCtx, GameDispatchCtx, gameReducer } from 'game/state';
-import type { GameDispatch } from 'game/state';
 import { EventDispatchCtx, eventListReducer } from 'event/state';
-import type { EventDispatch } from 'event/state';
 import * as server from 'server';
 
-function initialCookieCheck(dispatch: GameDispatch, eventDispatch: EventDispatch) {
-    const authMatch = document.cookie.match(/srAuth=[^.]+.([^.]+)/);
-    if (!authMatch) {
-        return;
-    }
-    const auth = JSON.parse(atob(authMatch[1]));
-    dispatch({
-        ty: "join",
-        gameID: auth.gid,
-        player: { id: auth.pid, name: auth.pname },
-        players: {}
-    });
-    eventDispatch({
-        ty: "gameJoin", gameID: auth.gid
-    })
-    server.getPlayers().then(players => {
-        dispatch({
-            ty: "setPlayers", players
-        })
-    })
-}
-
 const AppPadding: StyledComponent<> = styled.div`
-    margin: 0 auto;
+    margin: 0px auto;
+    display: flex;
+    flex-direction: column;
 
     max-width: 95%;
-    @media all and (min-width: 768) {
-        max-width: 98%;
+    @media all and (min-width: 768px) {
+        flex-direction: row;
+        max-width: 98%; /* align items */
+    }
+`;
+
+const AppLeft: StyledComponent<> = styled.div`
+    /* Phones: padding between roll & history */
+    padding: 5px 0px;
+
+    /* Tablet+: roll history on right. */
+    @media all and (min-width: 768px) {
+        padding: 0px;
+        margin-right: 5px;
+        flex-grow: 3;
+    }
+`;
+
+const AppRight: StyledComponent<> = styled.div`
+    /* Phones: no padding needed. */
+
+    @media all and (min-width: 768px) {
+        flex-grow: 1;
     }
 `;
 
@@ -57,24 +56,32 @@ export default function App(props: {}) {
         setShowGameJoin((state) => !state);
     }
 
-    React.useEffect(() => initialCookieCheck(gameDispatch, eventDispatch), []);
+    React.useEffect(() => server.initialCookieCheck(gameDispatch, eventDispatch), []);
 
     // Page should be a flexbox.
     return (
-            <GameCtx.Provider value={game}>
-            <GameDispatchCtx.Provider value={gameDispatch}>
-                <SRHeader game={game}
-                          expanded={showGameJoin}
-                          onClick={joinGameClicked} />
-                {showGameJoin ? <JoinGamePrompt game={game} setShown={setShowGameJoin} /> : ''}
-                <AppPadding>
+        <GameCtx.Provider value={game}>
+        <GameDispatchCtx.Provider value={gameDispatch}>
+
+            <SRHeader game={game} expanded={showGameJoin} onClick={joinGameClicked} />
+            {showGameJoin ? <JoinGamePrompt game={game} setShown={setShowGameJoin} /> : ''}
+
+            <AppPadding>
+
                 <EventDispatchCtx.Provider value={eventDispatch}>
+
+                <AppLeft>
                     <RollDicePrompt game={game}
                                     dispatch={eventDispatch} />
+                </AppLeft>
+                <AppRight>
                     <EventHistory eventList={eventList} />
+                </AppRight>
                 </EventDispatchCtx.Provider>
-                </AppPadding>
-            </GameDispatchCtx.Provider>
-            </GameCtx.Provider>
+
+            </AppPadding>
+
+        </GameDispatchCtx.Provider>
+        </GameCtx.Provider>
     );
 }
